% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_df_cols.R
\name{coerce_df_cols}
\alias{coerce_df_cols}
\title{Coerce classes in dataframe based on reference dataframe}
\usage{
coerce_df_cols(truth, source)
}
\arguments{
\item{truth}{truth dataframe to base new classes on}

\item{source}{source dataframe whose clases should be updated}
}
\value{
source dataframe with classes coerced based on truth dataframe
}
\description{
Coerces column classes of a dataframe based on a ref dataframe
}
\examples{
\dontrun{
if(interactive()){
 df1 <- data.frame(a = sample(1:10, 5), b = as.character(1:5))
 df2 <- data.frame(a = as.character(sample(1:10, 5)), b = 1:5)
 coerce_df_cols(truth = df1, source = df2)
 }
}
}
\seealso{
\code{\link[janitor]{compare_df_cols}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}
 \code{\link[purrr]{map2}}
 \code{\link[tibble]{as_tibble}}
}
