% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hex_sticker.R
\name{gen_hex_sticker}
\alias{gen_hex_sticker}
\title{Generate a hex sticker from an image file}
\usage{
gen_hex_sticker(
  ref_image,
  file_name = "hex.png",
  pkg_name,
  git_name = NULL,
  ft = "bangers",
  ft1 = "rob",
  txt_main_color = "black",
  txt_repo_color = "black",
  txt_main_size = 55,
  txt_repo_size = 25
)
}
\arguments{
\item{ref_image}{file path to reference image}

\item{file_name}{output image file, Default: 'hex.png'}

\item{pkg_name}{package name to include in hex sticker}

\item{git_name}{github repo name to include in hex sticker, Default: NULL}

\item{ft}{main font, Default: 'bangers'}

\item{ft1}{annotation font, Default: 'rob'}

\item{txt_main_color}{package name text color, Default: 'black'}

\item{txt_repo_color}{repo text color, Default: 'black'}

\item{txt_main_size}{package name text size, Default: 55}

\item{txt_repo_size}{repo text size, Default: 25}
}
\value{
string with path to image file
}
\description{
Create and open a hex sticker using a reference image file
}
\details{
function based on based on http://gradientdescending.com/how-to-generate-a-hex-sticker-with-openai-and-cropcircles/
}
\examples{
\dontrun{
if(interactive()){
  ref_image <- "man/figures/DALL·E 2023-07-15 23.59.36 - Ukrainian cat wearing a beret and black turtleneck.png"
  file_name <- "man/figures/hex.png"
  pkg_name <- "psam"
  git_name <- "parmsam/psam"
  gen_hex_sticker(
     ref_image,
     file_name,
     pkg_name,
     git_name,
     txt_main_color = "white",
     txt_repo_color = "black"
  )
 }
}
}
\seealso{
\code{\link[sysfonts]{font_add_google}}
 \code{\link[showtext]{showtext_auto}}
 \code{\link[cropcircles]{crop}}
 \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{annotate}}, \code{\link[ggplot2]{lims}}, \code{\link[ggplot2]{ggtheme}}, \code{\link[ggplot2]{coord_fixed}}, \code{\link[ggplot2]{ggsave}}
 \code{\link[ggpath]{geom_from_path}}
 \code{\link[utils]{browseURL}}
}
