% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_book_clean_code.R
\name{get_book_clean_code}
\alias{get_book_clean_code}
\title{Get clean code from book repo}
\usage{
get_book_clean_code(repo_name, branch_name, folder = "examples")
}
\arguments{
\item{repo_name}{Github repository name with org or username included}

\item{branch_name}{Github repository brach name}

\item{folder}{folder to export files, defaults to 'examples' folder}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
get_book_clean_code() retrieves clean R code from book repository based on Rmd files
}
\examples{
\dontrun{
if(interactive()){
 # get_book_clean_code("tidymodels/TMwR", "main", "examples")
 }
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{stop_for_status}}
 \code{\link[glue]{glue}}
 \code{\link[stringr]{str_subset}}, \code{\link[stringr]{str_replace}}
 \code{\link[purrr]{map2}}
 \code{\link[knitr]{knit}}
}
