% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gh_rscripts.R
\name{get_gh_rscripts}
\alias{get_gh_rscripts}
\title{Get list of R script Github URLs from Repo}
\usage{
get_gh_rscripts(repo_name, branch_name, filter_pattern = "\\\\.R$")
}
\arguments{
\item{repo_name}{Github repository name with org or username included}

\item{branch_name}{Github repository brach name}

\item{filter_pattern}{regex pattern to filter file list by, '*R$' by default for R files}

\item{folder}{folder to export files, defaults to 'examples' folder}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
get_gh_rscripts() retrieves sourceable R code URLs from a Github repo
}
\examples{
\dontrun{
if(interactive()){
 # get_gh_rscripts("tidymodels/TMwR", "main")
 }
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{stop_for_status}}
 \code{\link[glue]{glue}}
 \code{\link[stringr]{str_subset}}, \code{\link[stringr]{str_replace}}
}
